Moved Paths to app/Paths
